substituido para classe freight no caso de uso e movido calculo do frete para resource
