--- src/main/kotlin/com/montebruni/sales/domain/entity/freightCalculator/FreightCalculatorHandler.kt
@@ -1,7 +1,0 @@
-package com.montebruni.sales.domain.entity.freightCalculator
-
-import com.montebruni.sales.domain.entity.freightCalculator.input.FreightCalculatorInput
-
-interface FreightCalculatorHandler {
-    fun calculate(input: FreightCalculatorInput) : Double
-}
