+++ src/main/kotlin/com/montebruni/sales/resource/calculator/freightCalculator/FreightCalculatorHandler.kt
@@ -1,0 +1,7 @@
+package com.montebruni.sales.resource.calculator.freightCalculator
+
+import com.montebruni.sales.resource.calculator.freightCalculator.input.FreightCalculatorInput
+
+interface FreightCalculatorHandler {
+    fun calculate(input: FreightCalculatorInput) : Double
+}
