+++ src/main/kotlin/com/montebruni/sales/usecase/CalculateFreight.kt
@@ -1,21 +1,23 @@
 package com.montebruni.sales.usecase
 
-import com.montebruni.sales.domain.entity.freightCalculator.FreightCalculator
-import com.montebruni.sales.domain.entity.freightCalculator.input.FreightCalculatorInput
+import com.montebruni.sales.domain.entity.Freight
+import com.montebruni.sales.domain.port.FreightCalculator
+import com.montebruni.sales.extensions.toPositiveDouble
 import com.montebruni.sales.usecase.input.CalculateFreightInput
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 
 @Service
 class CalculateFreight(
-    private val freightCalculator: FreightCalculator
+    @Autowired private val freightCalculator: FreightCalculator
 ){
     fun execute(input: CalculateFreightInput): Double = input.items.sumOf {
         it.quantity * freightCalculator.calculate(
-            FreightCalculatorInput(
-                height = it.height.value,
-                width = it.width.value,
-                length = it.length.value,
-                weight = it.weight.value
+            Freight(
+                height = it.height.value.toPositiveDouble(),
+                width = it.width.value.toPositiveDouble(),
+                length = it.length.value.toPositiveDouble(),
+                weight = it.weight.value.toPositiveDouble()
             )
         )
     }
