+++ src/test/kotlin/com/montebruni/sales/domain/freightCalculator/handle/DefaultFreightCalculatorTest.kt
@@ -1,8 +1,8 @@
 package com.montebruni.sales.domain.freightCalculator.handle
 
 import com.montebruni.sales.common.UnitTests
-import com.montebruni.sales.domain.entity.freightCalculator.handle.DefaultFreightCalculator
-import com.montebruni.sales.fixture.domain.freightCalculator.createFreightCalculatorInput
+import com.montebruni.sales.resource.calculator.freightCalculator.handle.DefaultFreightCalculator
+import com.montebruni.sales.fixture.resource.calculator.freightCalculator.createFreightCalculatorInput
 import io.mockk.impl.annotations.InjectMockKs
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.params.ParameterizedTest
