+++ src/test/kotlin/com/montebruni/sales/domain/freightCalculator/handle/VolumeFreightCalculatorTest.kt
@@ -1,10 +1,10 @@
 package com.montebruni.sales.domain.freightCalculator.handle
 
 import com.montebruni.sales.common.UnitTests
-import com.montebruni.sales.domain.entity.freightCalculator.handle.DensityFreightCalculator
-import com.montebruni.sales.domain.entity.freightCalculator.input.FreightCalculatorInput
-import com.montebruni.sales.domain.entity.freightCalculator.handle.VolumeFreightCalculator
-import com.montebruni.sales.fixture.domain.freightCalculator.createFreightCalculatorInput
+import com.montebruni.sales.resource.calculator.freightCalculator.handle.DensityFreightCalculator
+import com.montebruni.sales.resource.calculator.freightCalculator.input.FreightCalculatorInput
+import com.montebruni.sales.resource.calculator.freightCalculator.handle.VolumeFreightCalculator
+import com.montebruni.sales.fixture.resource.calculator.freightCalculator.createFreightCalculatorInput
 import io.mockk.every
 import io.mockk.impl.annotations.InjectMockKs
 import io.mockk.impl.annotations.MockK
